backend/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
pydantic==2.12.5
httpx==0.28.1
pdfplumber==0.11.8
beautifulsoup4==4.14.3
openai==2.29.0
python-multipart==0.0.20
sse-starlette==3.3.0
python-dotenv==1.2.1
